+++ glean-core/ios/GleanTests/Metrics/ObjectMetricTests.swift
@@ -5,29 +5,81 @@
 @testable import Glean
 import XCTest
 
-struct BalloonsObjectItem: Decodable, Equatable, ObjectSerialize {
+struct BalloonsObjectItem: Codable, Equatable {
     var colour: String?
     var diameter: Int64?
+    var anotherValue: Bool?
+
+    enum CodingKeys: String, CodingKey {
+        case colour = "colour"
+        case diameter = "diameter"
+        case anotherValue = "another_value"
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        if let colour = self.colour {
+            try container.encode(colour, forKey: .colour)
+        }
+        if let diameter = self.diameter {
+            try container.encode(diameter, forKey: .diameter)
+        }
+        if let anotherValue = self.anotherValue {
+            try container.encode(anotherValue, forKey: .anotherValue)
+        }
+    }
+}
+typealias BalloonsObject = [BalloonsObjectItem]
+
+// generated from
+//
+// ```
+// structure:
+//   type: object
+//   properties:
+//     key1:
+//       type: string
+//     another_value:
+//       type: number
+//     sub_array:
+//       type: array
+//       items:
+//          type: number
+// ```
+struct ToplevelObjectObject: Codable, Equatable, ObjectSerialize {
+    var key1: String?
+    var anotherValue: Int64?
+    var subArray: ToplevelObjectObjectSubArray = []
+
+    enum CodingKeys: String, CodingKey {
+        case key1 = "key1"
+        case anotherValue = "another_value"
+        case subArray = "sub_array"
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        if let key1 = self.key1 {
+            try container.encode(key1, forKey: .key1)
+        }
+        if let anotherValue = self.anotherValue {
+            try container.encode(anotherValue, forKey: .anotherValue)
+        }
+        if subArray.count > 0 {
+            let subArray = self.subArray
+            try container.encode(subArray, forKey: .subArray)
+        }
+    }
 
     func intoSerializedObject() -> String {
-        var data: [String] = []
-        if let val = self.colour {
-            var elem = "\"colour\":"
-            elem.append(val.intoSerializedObject())
-            data.append(elem)
-        }
-        if let val = self.diameter {
-            var elem = "\"diameter\":"
-            elem.append(val.intoSerializedObject())
-            data.append(elem)
-        }
-        let obj = data.joined(separator: ",")
-        let json = "{" + obj + "}"
+        let jsonEncoder = JSONEncoder()
+        let jsonData = try! jsonEncoder.encode(self)
+        let json = String(data: jsonData, encoding: String.Encoding.utf8)!
         return json
     }
 }
-
-typealias BalloonsObject = [BalloonsObjectItem]
+typealias ToplevelObjectObjectSubArray = [ToplevelObjectObjectSubArrayItem]
+typealias ToplevelObjectObjectSubArrayItem = Int64
 
 class ObjectMetricTypeTests: XCTestCase {
     override func setUp() {
@@ -131,5 +183,55 @@ class ObjectMetricTypeTests: XCTestCase {
         snapshot = metric.testGetValue("store2")!
         XCTAssertEqual(2, snapshot.count)
         XCTAssertEqual(expected, snapshot)
+    }
+
+     func testObjectDecodesFromSnakeCase() {
+         let metric = ObjectMetricType<BalloonsObject>(CommonMetricData(
+             category: "test",
+             name: "balloon",
+             sendInPings: ["store1"],
+             lifetime: .ping,
+             disabled: false
+         ))
+
+         XCTAssertNil(metric.testGetValue())
+
+         var balloons: BalloonsObject = []
+         balloons.append(BalloonsObjectItem(colour: "red", diameter: 5, anotherValue: true))
+         balloons.append(BalloonsObjectItem(colour: "green", anotherValue: false))
+         metric.set(balloons)
+
+         let snapshot = metric.testGetValue()!
+         XCTAssertNotNil(snapshot)
+         XCTAssertEqual(2, snapshot.count)
+
+         XCTAssertEqual(snapshot[0].colour, "red")
+         XCTAssertEqual(snapshot[0].diameter, 5)
+         XCTAssertEqual(snapshot[0].anotherValue, true)
+         XCTAssertEqual(snapshot[1].colour, "green")
+         XCTAssertNil(snapshot[1].diameter)
+         XCTAssertEqual(snapshot[1].anotherValue, false)
+     }
+
+    func testObjectWithStructureOnToplevel() {
+        let metric = ObjectMetricType<ToplevelObjectObject>(CommonMetricData(
+            category: "test",
+            name: "toplevel_object",
+            sendInPings: ["store1"],
+            lifetime: .ping,
+            disabled: false
+        ))
+
+        XCTAssertNil(metric.testGetValue())
+
+        let obj = ToplevelObjectObject(key1: "test", anotherValue: 3, subArray: [1, 2, 3])
+        metric.set(obj)
+
+        let snapshot = metric.testGetValue()!
+        XCTAssertNotNil(snapshot)
+
+        XCTAssertEqual("test", snapshot.key1)
+        XCTAssertEqual(3, snapshot.anotherValue)
+        XCTAssertEqual([1, 2, 3], snapshot.subArray)
     }
 }